respond to PR comments; move classes + test suite into dynamodb package
